add unit testing in README
