+++ src/engine/minimax/tictactoe/mod.rs
@@ -1,3 +1,3 @@
 pub mod evaluation; 
 pub mod maximizer; 
-pub mod test_ttt_minimax; 
+pub mod test_minimax;
