+++ src/engine/mod.rs
@@ -1,2 +1,1 @@
 pub mod minimax;
-pub mod minimax_test;
